% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undid_stage_three.r
\name{coef.UnDiDObj}
\alias{coef.UnDiDObj}
\title{Extract coefficients from \code{UnDiDObj}}
\usage{
\method{coef}{UnDiDObj}(object, level = c("agg", "sub"), ...)
}
\arguments{
\item{object}{A \code{UnDiDObj} object}

\item{level}{Specify either \code{"agg"} or \code{"sub"} to view the aggregate
or sub-aggregate results.}

\item{...}{other arguments}
}
\value{
A data frame of coefficient estimates
}
\description{
Extract coefficients from \code{UnDiDObj}
}
