% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undidR_package_variables.r
\name{.populate_undid_env}
\alias{.populate_undid_env}
\title{Populate the internal .undid_env environment}
\usage{
.populate_undid_env(env)
}
\arguments{
\item{env}{An environment to populate with package variables.}
}
\description{
This function populates the .undid_env environment with pre-defined
variables for use within the package. It is called during .onLoad().
}
\keyword{internal}
