% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undid_stage_three.r
\name{print.UnDiDObj}
\alias{print.UnDiDObj}
\title{Print method for \code{UnDiDObj}}
\usage{
\method{print}{UnDiDObj}(x, level = c("agg", "sub"), ...)
}
\arguments{
\item{x}{A \code{UnDiDObj} object.}

\item{level}{Specify either \code{"agg"} or \code{"sub"} to view the aggregate
or sub-aggregate results.}

\item{...}{other arguments}
}
\description{
Print method for \code{UnDiDObj}
}
