% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undid_stage_three.r
\name{summary.UnDiDObj}
\alias{summary.UnDiDObj}
\title{Summary method for \code{UnDiDObj}}
\usage{
\method{summary}{UnDiDObj}(object, level = c("all", "agg", "sub"), ...)
}
\arguments{
\item{object}{A \code{UnDiDObj} object}

\item{level}{Specify either \code{"agg"}, \code{"sub"}, or \code{"all"}, to view the
results at the aggregate level, the sub-aggregate level, or to view
both simultaneously.}

\item{...}{other arguments}
}
\description{
Summary method for \code{UnDiDObj}
}
