% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undid_date_formats.r
\name{undid_date_formats}
\alias{undid_date_formats}
\title{Shows valid date formats}
\usage{
undid_date_formats()
}
\value{
A list of valid date formats
}
\description{
The \code{undid_date_formats} function returns a list of all valid date formats
that can be used within the package
}
\examples{
undid_date_formats()
}
