% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_diff_df.r
\name{create_diff_df}
\alias{create_diff_df}
\title{Creates the \code{empty_diff_df.csv}}
\usage{
create_diff_df(
  init_filepath,
  date_format,
  freq,
  covariates = FALSE,
  freq_multiplier = FALSE,
  weights = "standard",
  filename = "empty_diff_df.csv",
  filepath = tempdir()
)
}
\arguments{
\item{init_filepath}{A string filepath to the \code{init.csv}.}

\item{date_format}{A string specifying the date format used in the
\code{init.csv}. Call \code{undid_date_formats()} to see a list of valid date formats.}

\item{freq}{A string indicating the length of the time periods to be used
when computing the differences in mean outcomes between periods at each silo.}

\item{covariates}{A vector of strings specifying covariates to be considered
at each silo. If \code{FALSE} uses covariates from the \code{init.csv}.
Defaults to \code{FALSE}.}

\item{freq_multiplier}{A numeric value or \code{FALSE}. Specify if the frequency
should be multiplied by a non-zero integer. For example, to consider two year
periods, set \verb{freq = "yearly", freq_multiplier = 2}. Defaults to \code{FAlSE}.}

\item{weights}{A string indicating the type of weighting to use in the
case of common adoption. Defaults to \code{"standard"}. The \code{"standard"} weight
is calculated as \eqn{w_s = \frac{N_s^{\text{post}}}{N_s^{\text{post}} + N_s^{\text{pre}}}}.}

\item{filename}{A string filename for the created .csv file. Defaults to
\code{empty_diff_df.csv}}

\item{filepath}{Filepath to save the .csv file. Defaults to \code{tempdir()}.}
}
\value{
The \code{empty_diff_df.csv} is created in the working directory.
Its file path is printed and a dataframe of the \code{empty_diff_df.csv}
is returned.
}
\description{
Creates the \code{empty_diff_df.csv} which lists all of the differences that
need to calculated at each silo in order to compute the aggregate ATT.
The \code{empty_diff_df.csv} is then sent out to each silo to be filled out.
}
\details{
Ensure that dates in the \code{init.csv} are entered consistently
in the same date format. Call \code{undid_date_formats()} to see a list of valid
date formats. Covariates specified when calling \code{create_diff_df} will
override any covariates specified in the \code{init.csv}.
}
\examples{
# Create a temporary `init.csv` file for this example
init_df <- data.frame(
  silo_name = c("71", "73"),
  start_time = c("1989", "1989"),
  end_time = c("2000", "2000"),
  treatment_time = c("1991", "control")
)
init_filepath <- tempfile(fileext = ".csv")
write.table(init_df, init_filepath, sep = ",", row.names = FALSE,
            col.names = TRUE, quote = FALSE)

# Example call to create_diff_df
create_diff_df(
  init_filepath = init_filepath,
  date_format = "yyyy",
  freq = "yearly"
)
unlink(file.path(tempdir(), c("init.csv", "empty_diff_df.csv")))
}
