% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undid_stage_three.r
\name{plot.UnDiDObj}
\alias{plot.UnDiDObj}
\title{Plot method for \code{UnDiDObj}}
\usage{
\method{plot}{UnDiDObj}(x, event = FALSE, event_window = NULL, ci = 0.95, ...)
}
\arguments{
\item{x}{A \code{UnDiDObj} object}

\item{event}{Logical. If \code{TRUE}, creates an event study plot. If
\code{FALSE} (default), creates a parallel trends plot.}

\item{event_window}{Numeric vector of length 2 specifying the event window
as c(start, end). Default is \code{NULL} (use all available periods).}

\item{ci}{Numeric between 0 and 1 specifying confidence level.
Default is 0.95.}

\item{...}{other arguments passed to plot}
}
\description{
Plot method for \code{UnDiDObj}
}
