% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undidR-package.R
\docType{package}
\name{undidR-package}
\alias{undidR}
\alias{undidR-package}
\title{undidR: Difference-In-Differences with Unpoolable Data}
\description{
A method for estimating difference-in-differences with unpoolable data. Allows for covariates, multiple groups, and staggered adoption.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ebjamieson97/undidR}
  \item Report bugs at \url{https://github.com/ebjamieson97/undidR/issues}
}

}
\author{
\strong{Maintainer}: Eric Jamieson \email{ericbrucejamieson@gmail.com}

}
\keyword{internal}
