% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/undidR-package.R
\docType{package}
\name{undidR-package}
\alias{undidR}
\alias{undidR-package}
\title{undidR: Difference-in-Differences with Unpoolable Data}
\description{
A framework for estimating difference-in-differences with unpoolable data, based on Karim, Webb, Austin, and Strumpf (2024) \href{https://arxiv.org/abs/2403.15910v2}{arXiv:2403.15910v2}. Supports common or staggered adoption, multiple groups, and the inclusion of covariates. Also computes p-values for the aggregate average treatment effect on the treated (ATT) via the randomization inference procedure described in MacKinnon and Webb (2020) \doi{10.1016/j.jeconom.2020.04.024}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ebjamieson97/undidR}
  \item Report bugs at \url{https://github.com/ebjamieson97/undidR/issues}
}

}
\author{
\strong{Maintainer}: Eric Jamieson \email{ericbrucejamieson@gmail.com} [copyright holder]

}
\keyword{internal}
